third_party/blink/web_tests/external/wpt/css/css-properties-values-api/at-property-shadow-expected.txt

This is a testharness.js-based test.
[FAIL] @property rules in shadow trees should be globally registered
  assert_equals: expected "2px" but got "calc(1px + 1px)"
Harness: the test ran to completion.
